Change AudioController's public API from samples to milliseconds
The sample rate of the currently open audio is not something that things
which do not interact with the raw audio data should have to care about,
or even know about.

Originally committed to SVN as r6426.

+++ aegisub/src/audio_rendering_style.h
@@ -58,8 +58,8 @@ protected:
 	~AudioRenderingStyleRanges() { }
 public:
 	/// @brief Add a range to the line
-	/// @param start First sample index in range
-	/// @param end   One past last sample index in range
+	/// @param start First milisecond in range
+	/// @param end   One past last milisecond in range
 	/// @param style Style of the range added
-	virtual void AddRange(int64_t start, int64_t end, AudioRenderingStyle style) = 0;
+	virtual void AddRange(int start, int end, AudioRenderingStyle style) = 0;
 };
